refactor: 외부영역에 발생했을 때 실행될 커스텀 훅 생성
