+++ src/hooks/useClickOutside.tsx
@@ -1,0 +1,26 @@
+import { useEffect } from 'react';
+
+const useClickOutside = (
+  refs: React.RefObject<HTMLElement>[],
+  callback: () => void,
+) => {
+  useEffect(() => {
+    function handleClickOutside(event: MouseEvent): void {
+      const isOutside = refs.every(
+        (ref) => ref.current && !ref.current.contains(event.target as Node),
+      );
+
+      if (isOutside) {
+        callback();
+      }
+    }
+
+    document.addEventListener('mousedown', handleClickOutside);
+
+    return () => {
+      document.removeEventListener('mousedown', handleClickOutside);
+    };
+  }, [refs, callback]);
+};
+
+export default useClickOutside;
